Created Entity, ImageAnimation, MouseManager, Button

+++ src/main/java/com/xcodiq/taterunner/util/editor/ImageEditor.java
@@ -4,5 +4,5 @@ import java.awt.*;
 
 public interface ImageEditor {
 
-	void drawImage(int x, int y, Image image);
+	void drawImage(double x, double y, Image image);
 }
